change of price datatype

+++ Application/Commands/Items/UpdateItems/UpdateItemCommand.cs
@@ -1,5 +1,5 @@
 ﻿using MediatR;
 
 namespace Application.Commands.Items.UpdateItems;
-public record UpdateItemCommand(int Id, string Title, string Description, int Price, bool IsActive): IRequest<Unit>;
+public record UpdateItemCommand(int Id, string Title, string Description, decimal Price, bool IsActive): IRequest<Unit>;
